+++ src/users_groups_lib/__init__.py
@@ -1,4 +1,5 @@
 """Exposed users_groups_lib classes and methods."""
-from users_groups_lib.errors import UserPermissionError, UserExistError, GroupNotExistError
+from users_groups_lib.errors import (UserPermissionError, UserExistError, UserInUseError, UserNotExistError,
+                                     GroupNotExistError)
 from users_groups_lib.entities import User
 from users_groups_lib.manager import UserManager