+++ src/users_groups_lib/errors/user/__init__.py
@@ -2,3 +2,4 @@
 from users_groups_lib.errors.user.user_exist_error import UserExistError
 from users_groups_lib.errors.user.user_permission_error import UserPermissionError
 from users_groups_lib.errors.user.user_not_exist_error import UserNotExistError
+from users_groups_lib.errors.user.user_in_use_error import UserInUseError